backend/requirements.txt
fastapi
uvicorn
pypdf
python-multipart
tiktoken
sentence-transformers
torch
numpy
faiss-cpu
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
bcrypt==4.0.1
python-jose
email-validator
python-dotenv
openai
